Classify initial Cloud test blockers by filename
